fix LogDownloadController: id is uint16

+++ src/AnalyzeView/LogDownloadController.cc
@@ -453,7 +453,7 @@ LogDownloadController::_findMissingData()
 
 //----------------------------------------------------------------------------------------
 void
-LogDownloadController::_requestLogData(uint8_t id, uint32_t offset, uint32_t count)
+LogDownloadController::_requestLogData(uint16_t id, uint32_t offset, uint32_t count)
 {
     if(_vehicle) {
         //-- APM "Fix"
